fix(compact): emit schema-valid tool-result messages and inline summarizer
The history-compaction summarizer built tool-result ModelMessages without
toolName and with a raw output value, casting each message `as ModelMessage`
to bypass the type check. The AI SDK rejects that shape, so generateText
threw, the summarizer caught it, and compaction silently fell back to the
deterministic excerpt draft — the user got an excerpt, not the LLM summary
they triggered /compact for.

- Extract toolResultOutput / jsonValue / AiSdkToolResultOutput into a shared
  ai-sdk-tool-output module, reused by the conversation backend and the
  summarizer so both emit identical, schema-valid tool-result content.
- Summarizer tool-result now carries toolName and the {type,value} output
  wrapper; drop the `as ModelMessage` casts and split text on role so each
  push matches exactly one ModelMessage arm.
- Inline createAiSdkConversationSummarizer into buildLlmHistorySummarizer
  (resolveModel + generateText? + maxOutputTokens?). Remove the thin wrapper
  module/type/barrel export and its pass-through test; update CLI + desktop
  callers.
- Regression test: a tool-bearing conversation produces schema-valid
  tool-result messages (toolName + wrapped output) and does not fall back.
