+++ packages/runtime/src/__tests__/history-compact-summarizer.test.ts
@@ -1,5 +1,5 @@
 /**
- * Tests for buildLlmHistorySummarizer — the host-supplied LLM summary that
+ * Tests for buildLlmHistorySummarizer — the AI-SDK-backed LLM summary that
  * replaces the deterministic excerpt draft when wiring injects it.
  *
  * Run: `npm --workspace @maka/runtime run test`
@@ -8,7 +8,7 @@ import { describe, test } from 'node:test';
 import { expect } from '../test-helpers.js';
 import type { RuntimeEvent, RuntimeEventContent } from '@maka/core/runtime-event';
 import type { HistoryCompactWriteInput } from '../ai-sdk-backend.js';
-import { buildLlmHistorySummarizer, type LlmConversationSummarizer } from '../history-compact-summarizer.js';
+import { buildLlmHistorySummarizer, type AiSdkGenerateTextLike } from '../history-compact-summarizer.js';
 
 const ts = 1_700_000_000_000;
 let __seq = 0;
@@ -45,14 +45,14 @@ function inputWith(events: RuntimeEvent[], abortSignal?: AbortSignal): HistoryCo
 }
 
 describe('buildLlmHistorySummarizer', () => {
-  test('returns the LLM summary and sends the tool-bearing conversation to the summarizer', async () => {
+  test('returns the LLM summary and sends the tool-bearing conversation to generateText', async () => {
     const seen: Array<{ system: string; messages: unknown[] }> = [];
-    const summarizeConversation: LlmConversationSummarizer = async (req) => {
-      seen.push(req);
-      return '## Goal\n做到 X';
+    const generateText: AiSdkGenerateTextLike = async (opts) => {
+      seen.push(opts);
+      return { text: '## Goal\n做到 X' };
     };
 
-    const summarize = buildLlmHistorySummarizer({ summarizeConversation });
+    const summarize = buildLlmHistorySummarizer({ resolveModel: () => 'fake-model', generateText });
 
     const events: RuntimeEvent[] = [
       ev({ role: 'user', author: 'user', content: { kind: 'text', text: '读 package.json' } }),
@@ -79,11 +79,49 @@ describe('buildLlmHistorySummarizer', () => {
     expect(serialized).toContain('maka');
   });
 
-  test('fail-open: returns undefined when the summarizer throws, so runtime falls back to the draft', async () => {
-    const summarizeConversation: LlmConversationSummarizer = async () => {
+  test('produces schema-valid tool-result messages (toolName + wrapped output) and does not fall back', async () => {
+    const seen: Array<{ messages: unknown[] }> = [];
+    const generateText: AiSdkGenerateTextLike = async (opts) => {
+      seen.push(opts);
+      return { text: '## Goal\nX' };
+    };
+    const summarize = buildLlmHistorySummarizer({ resolveModel: () => 'fake-model', generateText });
+
+    const events: RuntimeEvent[] = [
+      ev({ role: 'user', author: 'user', content: { kind: 'text', text: '读 package.json' } }),
+      ev({
+        role: 'model',
+        author: 'agent',
+        content: { kind: 'function_call', id: 'fc1', name: 'read', args: { path: 'package.json' } },
+      }),
+      ev({
+        role: 'tool',
+        author: 'tool',
+        content: { kind: 'function_response', id: 'fc1', name: 'read', result: { name: 'maka' } },
+      }),
+      ev({ role: 'model', author: 'agent', content: { kind: 'text', text: 'ok' } }),
+    ];
+
+    const result = await summarize(inputWith(events));
+    expect(result).toBe('## Goal\nX');
+
+    const messages = seen[0]!.messages as Array<{
+      role: string;
+      content: Array<{ type: string; toolName?: string; output?: unknown }>;
+    }>;
+    const toolPart = messages.find((m) => m.role === 'tool')!.content[0]!;
+    expect(toolPart.type).toBe('tool-result');
+    // toolName must be present (AI SDK v6 tool-result content requires it)
+    expect(toolPart.toolName).toBe('read');
+    // output must be the {type, value} wrapper, not the raw result object
+    expect(toolPart.output).toEqual({ type: 'json', value: { name: 'maka' } });
+  });
+
+  test('fail-open: returns undefined when generateText throws, so runtime falls back to the draft', async () => {
+    const generateText: AiSdkGenerateTextLike = async () => {
       throw new Error('model down');
     };
-    const summarize = buildLlmHistorySummarizer({ summarizeConversation });
+    const summarize = buildLlmHistorySummarizer({ resolveModel: () => 'fake-model', generateText });
 
     const result = await summarize(
       inputWith([ev({ role: 'user', author: 'user', content: { kind: 'text', text: 'hi' } })]),
@@ -92,13 +130,13 @@ describe('buildLlmHistorySummarizer', () => {
     expect(result).toBe(undefined);
   });
 
-  test('returns undefined without calling the summarizer when there are no events to summarize', async () => {
+  test('returns undefined without calling generateText when there are no events to summarize', async () => {
     let called = false;
-    const summarizeConversation: LlmConversationSummarizer = async () => {
+    const generateText: AiSdkGenerateTextLike = async () => {
       called = true;
-      return 'should not reach';
+      return { text: 'should not reach' };
     };
-    const summarize = buildLlmHistorySummarizer({ summarizeConversation });
+    const summarize = buildLlmHistorySummarizer({ resolveModel: () => 'fake-model', generateText });
 
     const result = await summarize(inputWith([]));
