+++ packages/runtime/src/history-compact-summarizer.ts
@@ -1,19 +1,27 @@
 import type { ModelMessage } from 'ai';
 import { buildRuntimeEventModelReplayPlan } from './model-history.js';
+import { toolResultOutput } from './ai-sdk-tool-output.js';
 import type { HistoryCompactWriteInput } from './ai-sdk-backend.js';
 
-export interface LlmConversationSummarizerInput {
+export interface AiSdkGenerateTextOptions {
+  model: unknown;
   system: string;
   messages: ModelMessage[];
+  maxOutputTokens?: number;
   abortSignal?: AbortSignal;
 }
 
-export type LlmConversationSummarizer = (
-  input: LlmConversationSummarizerInput,
-) => Promise<string>;
+export type AiSdkGenerateTextLike = (
+  options: AiSdkGenerateTextOptions,
+) => Promise<{ text: string }>;
 
 export interface BuildLlmHistorySummarizerOptions {
-  summarizeConversation: LlmConversationSummarizer;
+  /** Resolve the AI SDK model used for summarization. Reuses the session model. */
+  resolveModel: () => unknown;
+  /** Optional cap on the generated summary length. */
+  maxOutputTokens?: number;
+  /** Injectable `generateText` for tests; defaults to the real AI SDK export. */
+  generateText?: AiSdkGenerateTextLike;
 }
 
 // Conversation-summarization prompt (sectioned, modelled on pi/opencode):
@@ -54,11 +62,15 @@ export function buildLlmHistorySummarizer(options: BuildLlmHistorySummarizerOpti
     try {
       const plan = buildRuntimeEventModelReplayPlan(input.source.foldedRuntimeEvents);
       const messages = replayPlanItemsToModelMessages(plan.items);
-      return await options.summarizeConversation({
+      const generateText = options.generateText ?? (await loadAiSdkGenerateText());
+      const result = await generateText({
+        model: options.resolveModel(),
         system: SUMMARIZATION_SYSTEM_PROMPT,
         messages,
+        ...(options.maxOutputTokens !== undefined ? { maxOutputTokens: options.maxOutputTokens } : {}),
         ...(input.abortSignal ? { abortSignal: input.abortSignal } : {}),
       });
+      return result.text;
     } catch {
       // Fail-open: a summarizer failure returns undefined so the runtime
       // keeps the deterministic draft summary instead of aborting the compact.
@@ -67,28 +79,39 @@ export function buildLlmHistorySummarizer(options: BuildLlmHistorySummarizerOpti
   };
 }
 
+async function loadAiSdkGenerateText(): Promise<AiSdkGenerateTextLike> {
+  const ai = await import('ai').catch((err) => {
+    throw new Error(`Failed to load 'ai' package for history summarization. Run \`npm install ai\`. Inner: ${(err as Error).message}`);
+  });
+  const { generateText } = ai as { generateText: AiSdkGenerateTextLike };
+  return generateText;
+}
+
 type ReplayPlanItems = ReturnType<typeof buildRuntimeEventModelReplayPlan>['items'];
 
 function replayPlanItemsToModelMessages(items: ReplayPlanItems): ModelMessage[] {
   const out: ModelMessage[] = [];
   for (const item of items) {
     if (item.kind === 'text') {
-      out.push({
-        role: item.role === 'user' ? 'user' : 'assistant',
-        content: [{ type: 'text', text: item.content }],
-      } as ModelMessage);
+      // Split on role so each push matches exactly one ModelMessage arm — no cast.
+      const textPart = { type: 'text' as const, text: item.content };
+      if (item.role === 'user') {
+        out.push({ role: 'user', content: [textPart] });
+      } else {
+        out.push({ role: 'assistant', content: [textPart] });
+      }
     } else if (item.kind === 'tool_call') {
       out.push({
         role: 'assistant',
         content: [
           { type: 'tool-call', toolCallId: item.toolCallId, toolName: item.toolName, input: item.input },
         ],
-      } as ModelMessage);
+      });
     } else if (item.kind === 'tool_result') {
       out.push({
         role: 'tool',
-        content: [{ type: 'tool-result', toolCallId: item.toolCallId, output: item.output }],
-      } as ModelMessage);
+        content: [{ type: 'tool-result', toolCallId: item.toolCallId, toolName: item.toolName, output: toolResultOutput(item.output, item.isError) }],
+      });
     }
     // thinking entries are intentionally skipped for summarization
   }
